backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
language_tool_python
google-generativeai
